% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmi_core.R
\name{BLUP_impute}
\alias{BLUP_impute}
\title{Impute missing covariate values using BLUP}
\usage{
BLUP_impute(
  IPD,
  AgD,
  AgD_EM_cols,
  IPD_EM_cols,
  Study_col,
  samp_sizes,
  AgD_Trt_cols,
  TE_col,
  IPD_Trt_col,
  SE_col,
  outcome_col,
  outcome_type = "binary"
)
}
\arguments{
\item{IPD}{An IPD dataset.}

\item{AgD}{An NMI-format AgD dataset.}

\item{AgD_EM_cols}{A vector of effect modifier column names in the AgD.}

\item{IPD_EM_cols}{A vector of effect modifier column names in the IPD.}

\item{Study_col}{Study identifier column name for the AgD.}

\item{samp_sizes}{A vector of sample sizes for the AgD studies.}

\item{AgD_Trt_cols}{A vector of treatment column names for the AgD.}

\item{TE_col}{Relative treatment effect column name for the AgD.}

\item{IPD_Trt_col}{Treatment column name for the IPD.}

\item{SE_col}{Standard error column name for the AgD.}

\item{outcome_col}{Outcome column name for the IPD.}

\item{outcome_type}{Type of outcome: "binary", "continuous", or "count".}
}
\value{
The enriched (imputed) data frame, with the summarised IPD data
concatenated at the bottom.
}
\description{
Enriches the NMI AgD table by imputing missing covariate values using
the Best Linear Unbiased Predictor (BLUP).
}
