% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{display_result_table}
\alias{display_result_table}
\title{Display results table with formatting}
\usage{
display_result_table(NMA_results, NMR_results, ML_NMR_results, NMI_results)
}
\arguments{
\item{NMA_results}{The output of \code{\link{result_table}} for an NMA run.}

\item{NMR_results}{The output of \code{\link{result_table}} for an NMR run.}

\item{ML_NMR_results}{The output of \code{\link{result_table}} for an ML-NMR run.}

\item{NMI_results}{The output of \code{\link{result_table}} for an NMI run.}
}
\value{
A formatted kableExtra table.
}
\description{
Creates a formatted table for displaying results from different methods.
}
