% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_AgD_NMI}
\alias{Example_AgD_NMI}
\title{Example Aggregate Data (AgD) for NMI}
\format{
A data frame with 8 variables:
\describe{
\item{Study}{Study identifier}
\item{Trt1}{First treatment in comparison}
\item{Trt2}{Second treatment in comparison}
\item{n}{Sample size (for some rows)}
\item{x1}{Proportion of patients with first effect modifier}
\item{x2}{Proportion of patients with second effect modifier}
\item{TE}{Treatment effect estimate}
\item{se}{Standard error of treatment effect}
}
}
\source{
Simulated data based on Harari et al. (2023) Network Meta-Interpolation paper
}
\usage{
Example_AgD_NMI
}
\description{
A dataset containing aggregate data in the format required for Network
Meta-Interpolation (NMI) analysis.
}
\references{
Harari et al. (2023) Network meta-interpolation: Effect modification
adjustment in network meta-analysis using subgroup analyses.
Journal of the Royal Statistical Society: Series A.
}
\keyword{datasets}
