% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_reporting.R
\name{generate_data_summary_tab}
\alias{generate_data_summary_tab}
\title{Generate simple placeholder tabs}
\usage{
generate_data_summary_tab(nmi_results, include_data_tables)
}
\description{
Generate simple placeholder tabs
}
