% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_reporting.R
\name{generate_html_body_close}
\alias{generate_html_body_close}
\title{Generate HTML Body Closing}
\usage{
generate_html_body_close()
}
\value{
HTML body closing string
}
\description{
Generate HTML Body Closing
}
