% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_reporting.R
\name{generate_html_body_open}
\alias{generate_html_body_open}
\title{Generate HTML Body Opening}
\usage{
generate_html_body_open()
}
\value{
HTML body opening string
}
\description{
Generate HTML Body Opening
}
