% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_reporting.R
\name{generate_html_header}
\alias{generate_html_header}
\title{Generate HTML Header}
\usage{
generate_html_header(title)
}
\arguments{
\item{title}{Report title}
}
\value{
HTML header string
}
\description{
Generate HTML Header
}
