% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_reporting.R
\name{generate_navigation_tabs}
\alias{generate_navigation_tabs}
\title{Generate Navigation Tabs}
\usage{
generate_navigation_tabs()
}
\value{
HTML navigation string
}
\description{
Generate Navigation Tabs
}
