% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_reporting.R
\name{generate_overview_tab}
\alias{generate_overview_tab}
\title{Generate Overview Tab}
\usage{
generate_overview_tab(nmi_results)
}
\arguments{
\item{nmi_results}{NMI results object}
}
\value{
HTML overview tab string
}
\description{
Generate Overview Tab
}
