% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_reporting.R
\name{generate_report_header}
\alias{generate_report_header}
\title{Generate Report Header}
\usage{
generate_report_header(title, author, organization, timestamp, nmi_results)
}
\arguments{
\item{title}{Report title}

\item{author}{Author name}

\item{organization}{Organization name}

\item{timestamp}{Report timestamp}

\item{nmi_results}{NMI results object}
}
\value{
HTML header string
}
\description{
Generate Report Header
}
