% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmi_core.R
\name{GLMLOGIT}
\alias{GLMLOGIT}
\title{Turning IPD into subgroup analyses}
\usage{
GLMLOGIT(data, EM_cols, Trt_col, outcome_col, outcome_type = "binary")
}
\arguments{
\item{data}{An IPD dataset.}

\item{EM_cols}{A vector of effect modifier column names.}

\item{Trt_col}{Treatment assignment column name.}

\item{outcome_col}{Outcome column name.}

\item{outcome_type}{Type of outcome: "binary", "continuous", or "count".}
}
\value{
A data frame consisting of 2\emph{p + 1 rows, the first of which contains
the TE and SE for the entire trial, and the remaining 2}p contain subgroup
analyses with missing EM values for all but one effect modifier.
}
\description{
Transforms individual patient data into subgroup analyses matching the data
format required for NMI with different outcome types and binary effect modifiers.
}
