% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_nmr_functions.R
\name{ML_NMR_Run_3D}
\alias{ML_NMR_Run_3D}
\title{Run ML-NMR for different outcome types with 3 effect modifiers}
\usage{
ML_NMR_Run_3D(
  ML_NMR_data,
  N_iter,
  N_chains,
  burnin,
  n_int,
  outcome_type = "binary"
)
}
\arguments{
\item{ML_NMR_data}{A list consisting of the following elements:
\describe{
\item{AgD}{Aggregate-level data.}
\item{IPD}{Individual patient-level data.}
}}

\item{N_iter}{Number of total iterations per chain (including burn in).}

\item{N_chains}{Number of Markov chains.}

\item{burnin}{Length of burn in, i.e. number of iterations to discard at the
beginning.}

\item{n_int}{Number of Monte Carlo integration points.}

\item{outcome_type}{Type of outcome: "binary", "continuous", or "count".}
}
\value{
A \code{\link[multinma]{stan_nma}} object.
}
\description{
Runs multilevel network meta-regression for different outcome types with
three effect modifiers using the multinma package.
}
