% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_nmr_functions.R
\name{ML_NMR_summary_2D}
\alias{ML_NMR_summary_2D}
\title{Extract treatment effects from ML-NMR run (2D)}
\usage{
ML_NMR_summary_2D(n_trts, ML_NMR_Fit, x_vect)
}
\arguments{
\item{n_trts}{Number of distinct treatments in the network.}

\item{ML_NMR_Fit}{A \code{\link[multinma]{stan_nma}} object. The output of
\code{\link{ML_NMR_Run_2D}}.}

\item{x_vect}{A vector of length two, consisting of the effect modifier
levels at which the ITC will be conducted.}
}
\value{
A data frame of posterior summaries for the treatment effects at
x_vect.
}
\description{
Summarizes posterior distributions of treatment effects from a multilevel
network meta-regression run with two effect modifiers.
}
