% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_functions.R
\name{NMA_Meta_Reg_run_3D}
\alias{NMA_Meta_Reg_run_3D}
\title{Run Network Meta-Regression with 3 effect modifiers}
\usage{
NMA_Meta_Reg_run_3D(dat, N_chains, N_iter, burnin)
}
\arguments{
\item{dat}{Aggregate level data.}

\item{N_chains}{Number of Markov chains.}

\item{N_iter}{Number of total iterations per chain (including burn in).}

\item{burnin}{Length of burn in, i.e. number of iterations to discard at the
beginning.}
}
\value{
A \code{\link[rstan]{stanfit}} object.
}
\description{
Runs a fixed effects network meta-regression with three effect modifiers
using Stan.
}
