% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_functions.R
\name{NMA_Metareg_summary_2D}
\alias{NMA_Metareg_summary_2D}
\title{Extract treatment effects from NMR run (2D)}
\usage{
NMA_Metareg_summary_2D(sim, x_vect)
}
\arguments{
\item{sim}{A \code{\link[rstan]{stanfit}} object. The output of
\code{\link{NMA_Meta_Reg_run_2D}}.}

\item{x_vect}{A vector of length two, consisting of the effect modifier
levels at which the ITC will be conducted.}
}
\value{
A data frame of posterior summaries for the treatment effects at x1
and x2.
}
\description{
Summarizes posterior distributions of treatment effects from a network
meta-regression run with two effect modifiers.
}
