% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_functions.R
\name{NMA_Metareg_summary_3D}
\alias{NMA_Metareg_summary_3D}
\title{Extract treatment effects from NMR run (3D)}
\usage{
NMA_Metareg_summary_3D(sim, x_vect)
}
\arguments{
\item{sim}{An \code{\link[R2OpenBUGS]{bugs}} object. The output of
\code{\link{NMA_Meta_Reg_run_3D}}.}

\item{x_vect}{A vector of length three, consisting of the effect modifier
levels at which the ITC will be conducted.}
}
\value{
A data frame of posterior summaries for the treatment effects at x1,
x2, and x3.
}
\description{
Summarizes posterior distributions of treatment effects from a network
meta-regression run with three effect modifiers.
}
