% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_functions.R
\name{NMA_NMI_summary}
\alias{NMA_NMI_summary}
\title{Extract treatment effects from NMA run}
\usage{
NMA_NMI_summary(sim)
}
\arguments{
\item{sim}{The output of \code{\link{NMA_run}} (cmdstanr format).}
}
\value{
A data frame of posterior summaries for the treatment effects.
}
\description{
Summarizes posterior distributions of treatment effects from an NMA run.
}
