% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{NMI_diagnostic_plot}
\alias{NMI_diagnostic_plot}
\title{NMI interpolation diagnostic plot}
\usage{
NMI_diagnostic_plot(NMI_object)
}
\arguments{
\item{NMI_object}{The output of \code{\link{NMI_interpolation}}.}
}
\value{
A ggplot object displaying the predicted TE estimates from the
NMI algorithm vs. the observed TE estimates.
}
\description{
Creates a diagnostic plot showing observed vs predicted treatment effects
to assess goodness of interpolation.
}
