% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{NMI_diagnostic_plotly}
\alias{NMI_diagnostic_plotly}
\title{Interactive NMI interpolation diagnostic plot}
\usage{
NMI_diagnostic_plotly(NMI_object)
}
\arguments{
\item{NMI_object}{The output of \code{\link{NMI_interpolation}}.}
}
\value{
An interactive plot displaying the predicted TE estimates from the
NMI algorithm vs. the observed TE estimates.
}
\description{
Creates an interactive diagnostic plot showing observed vs predicted treatment effects
to assess goodness of interpolation.
}
