% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmi_core.R
\name{NMI_interpolation}
\alias{NMI_interpolation}
\title{NMI interpolation pre-NMA}
\usage{
NMI_interpolation(
  IPD,
  AgD,
  x_vect,
  AgD_EM_cols,
  IPD_EM_cols,
  Study_col,
  samp_sizes,
  AgD_Trt_cols,
  TE_col,
  SE_col,
  IPD_Trt_col,
  outcome_col,
  outcome_type = "binary"
)
}
\arguments{
\item{IPD}{An IPD dataset.}

\item{AgD}{An NMI-format AgD dataset.}

\item{x_vect}{A vector consisting of the effect modifier values for ITC.}

\item{AgD_EM_cols}{A vector of effect modifier column names in the AgD.}

\item{IPD_EM_cols}{A vector of effect modifier column names in the IPD.}

\item{Study_col}{Study identifier column name for the AgD.}

\item{samp_sizes}{A vector of sample sizes for the AgD studies.}

\item{AgD_Trt_cols}{A vector of treatment column names for the AgD.}

\item{TE_col}{Relative treatment effect column name for the AgD.}

\item{SE_col}{Standard error column name for the AgD.}

\item{IPD_Trt_col}{Treatment column name for the IPD.}

\item{outcome_col}{Outcome column name for the IPD.}

\item{outcome_type}{Type of outcome: "binary", "continuous", or "count".}
}
\value{
A list consisting of three data frames:
\item{Imputed}{The imputed AgD - the output of \code{\link{BLUP_impute}}.}
\item{Final}{The final NMA-format data set, after interpolation at the new
effect modifier levels.}
\item{Diagnostics}{Observed and predicted TEs for goodness of interpolation
diagnostics.}
}
\description{
Interpolates treatment effect estimates and standard errors at new
effect modifier values using the Network Meta-Interpolation methodology.
}
