% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{result_forest_plot}
\alias{result_forest_plot}
\title{Forest plot for treatment effects comparison}
\usage{
result_forest_plot(
  NMA_summary,
  NMR_summary,
  ML_NMR_summ,
  NMI_summary,
  trt_effs = NULL
)
}
\arguments{
\item{NMA_summary}{The output of \code{\link{NMA_NMI_summary}} for an NMA run.}

\item{NMR_summary}{The output of \code{\link{NMA_Metareg_summary_2D}}.}

\item{ML_NMR_summ}{The output of \code{\link{ML_NMR_summary_2D}}.}

\item{NMI_summary}{The output of \code{\link{NMA_NMI_summary}} for an NMI run.}

\item{trt_effs}{A vector of the true treatment effect used to generate the
data. Default is NULL.}
}
\value{
A ggplot2 object.
}
\description{
Creates a forest plot comparing treatment effects from different methods.
}
