% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{result_table}
\alias{result_table}
\title{Format treatment effects table}
\usage{
result_table(summary_tab)
}
\arguments{
\item{summary_tab}{A data frame. The output of \code{\link{NMA_NMI_summary}},
\code{\link{NMA_Metareg_summary_2D}}, or \code{\link{ML_NMR_summary_2D}}.}
}
\value{
A single row data frame containing posterior medians and 95\% CrIs for
all treatment effects.
}
\description{
Formats posterior summaries into a nice table with credible intervals.
}
