% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr_functions.R
\name{stan_model_2D}
\alias{stan_model_2D}
\title{Stan model for fixed effects NMR with 2 effect modifiers}
\usage{
stan_model_2D()
}
\value{
Character string containing Stan model code
}
\description{
Stan model code for fixed effects network meta-regression with continuous
outcomes and two effect modifiers in a treatment-difference format.
}
